# Add buffer to speed up transaction streaming