--- a/ledger/participant-integration-api/src/main/scala/platform/store/dao/events/TransactionsReader.scala
+++ b/ledger/participant-integration-api/src/main/scala/platform/store/dao/events/TransactionsReader.scala
@@ -5,6 +5,7 @@ package com.daml.platform.store.dao.events
 
 import java.sql.Connection
 
+import akka.stream.OverflowStrategy
 import akka.stream.scaladsl.Source
 import akka.{Done, NotUsed}
 import com.daml
@@ -49,6 +50,10 @@ private[dao] final class TransactionsReader(
   private val sqlFunctions = SqlFunctions(dbType)
 
   private val logger = ContextualizedLogger.get(this.getClass)
+
+  // TransactionReader adds an Akka stream buffer at the end of all streaming queries.
+  // This significantly improves the performance of the transaction service.
+  private val outputStreamBufferSize = 128
 
   private def offsetFor(response: GetTransactionsResponse): Offset =
     ApiOffset.assertFromString(response.transactions.head.offset)
@@ -115,6 +120,7 @@ private[dao] final class TransactionsReader(
         val response = EventsTable.Entry.toGetTransactionsResponse(events)
         response.map(r => offsetFor(r) -> r)
       }
+      .buffer(outputStreamBufferSize, OverflowStrategy.backpressure)
       .wireTap(_ match {
         case (_, response) =>
           response.transactions.foreach(txn =>
@@ -201,6 +207,7 @@ private[dao] final class TransactionsReader(
         val response = EventsTable.Entry.toGetTransactionTreesResponse(events)
         response.map(r => offsetFor(r) -> r)
       }
+      .buffer(outputStreamBufferSize, OverflowStrategy.backpressure)
       .wireTap(_ match {
         case (_, response) =>
           response.transactions.foreach(txn =>
@@ -280,6 +287,7 @@ private[dao] final class TransactionsReader(
 
     groupContiguous(events)(by = _.transactionId)
       .mapConcat(EventsTable.Entry.toGetActiveContractsResponse)
+      .buffer(outputStreamBufferSize, OverflowStrategy.backpressure)
       .wireTap(response => {
         span.addEvent(
           com.daml.metrics.Event("contract", Map((SpanAttribute.Offset.key, response.offset)))