This PR fixes an issue that was causing the transaction service to become almost twice as slow. I still don't understand what exactly is going on, but this small PR restores the original performance. 

## Changelog

```
CHANGELOG_BEGIN
- [Ledger API Server] Fixed an issue that was causing a slowdown of the
  transaction service.
CHANGELOG_END
```

## Measuring the impact

Method:
- Generate an index database with huntsman (many thanks to Martino)
- Build ledger-on-memory from commit 9ed787cb
- Run ledger-on-memory in ledger-api-only mode, connect to the existing index database
- Run spider-mock to measure the speed of the transaction service
- Repeat with ledger-on-memory where the out-of-order fix was reverted

## Results

Reading 17009 transactions took:
```
mergeSubstreams():                 18.602 secs (before the out-of-order fix)
concatSubstreams():                29.007 secs (after the out-of-order fix)
statefulMapConcat():               28.656 secs (first attempt: code without substreams)
statefulMapConcat() + 2x buffer(): 18.767 secs (second attempt: add buffers)
concatSubstreams() + buffer():     18.981 secs (third attempt: smallest change)
```

Code for the `statefulMapConcat` versions: 
```
source
  .map(Some(_))
  .concat(Source.single(None))
  .statefulMapConcat(() => {
    var previousSegmentKey: K = null.asInstanceOf[K]
    val builder = new VectorBuilder[A]
    entryO => {
      entryO match {
        case Some(entry) =>
          val keyForEntry = by(entry)
          if (keyForEntry != previousSegmentKey) {
            previousSegmentKey = keyForEntry
            val result = builder.result()
            builder.clear()
            builder += entry
            List(result)
          } else {
            builder += entry
            List.empty
          }
        case None =>
          List(builder.result())
      }
    }
  })
```