@SamirTalwar-DA 2021-01-20
> Thank you so much for the very accurate analysis.

I second this. It was very helpful!