@stefanobaghino-da 2021-01-21
> @stefanobaghino-da @SamirTalwar-DA I have moved the `.buffer()` call to TransactionReader and have created a documented private val for the magic number. I felt the buffer call does not belong to the generic `groupContiguous` method. I have verified that the performance of the transaction stream is slightly better than before the fix:
> 
> ```
> mergeSubstreams():             16.7 secs (before the out-of-order fix)
> concatSubstreams():            26.6 secs (after the out-of-order fix)
> concatSubstreams() + buffer(): 16.4 secs (after this PR)
> ```

Given the hard-coded number it probably makes more sense like this. Do you know what `.mergeSubstreams` does?