@rautenrieth-da 2021-01-21
> Do you know what .mergeSubstreams does?

I don't, unfortunately. All I know is that concatSubstreams() is implemented as mergeSubstreams() with maxParallelism=1.